Split inference dependening upon build tool
