+++ pkg/rebuild/maven/infer_test.go
@@ -7,16 +7,10 @@ import (
 	"archive/zip"
 	"bytes"
 	"context"
-	"fmt"
 	"io"
-	"path"
 	"testing"
-	"time"
 
-	"github.com/go-git/go-billy/v5/memfs"
-	"github.com/go-git/go-git/v5"
-	"github.com/go-git/go-git/v5/plumbing/object"
-	"github.com/go-git/go-git/v5/storage/memory"
+	"github.com/google/oss-rebuild/internal/gitx/gitxtest"
 	"github.com/google/oss-rebuild/pkg/archive"
 	"github.com/google/oss-rebuild/pkg/rebuild/rebuild"
 	"github.com/google/oss-rebuild/pkg/registry/maven"
@@ -178,245 +172,71 @@ func TestGetClassFileMajorVersion(t *testing.T) {
 	}
 }
 
-func TestSourceRepositoryURLInference(t *testing.T) {
-	testCases := []struct {
-		name        string
-		pom         []PomXML
-		targetPom   rebuild.Target
-		expectedURL string
-		wantErr     bool
+func TestBuildToolInference(t *testing.T) {
+	for _, tc := range []struct {
+		name              string
+		repo              string
+		expectedBuildTool string
+		wantErr           bool
 	}{
 		{
-			name: "get SCM URL",
-			pom: []PomXML{
-				{
-					GroupID:    "org.example",
-					ArtifactID: "child",
-					VersionID:  "1.0.0",
-					SCMURL:     "https://github.com/example/child",
-					URL:        "https://example.com/child",
-				},
-			},
-			targetPom: rebuild.Target{
-				Package: "org.example:child",
-				Version: "1.0.0",
-			},
-			expectedURL: "https://github.com/example/child",
-			wantErr:     false,
+			name: "pom.xml present",
+			repo: `
+            commits:
+              - id: initial-commit
+                files:
+                  pom.xml: |
+                      <project></project>`,
+			expectedBuildTool: mavenBuildTool,
+			wantErr:           false,
 		},
 		{
-			name: "get URL",
-			pom: []PomXML{
-				{
-					GroupID:    "com.example",
-					ArtifactID: "child",
-					VersionID:  "1.0.0",
-					URL:        "https://example.com/child",
-				},
-			},
-			targetPom: rebuild.Target{
-				Package: "com.example:child",
-				Version: "1.0.0",
-			},
-			expectedURL: "https://example.com/child",
-			wantErr:     false,
+			name: "pom.xml absent",
+			repo: `
+            commits:
+              - id: initial-commit
+                files:
+                  README.md: |
+                      # Sample Project`,
+			expectedBuildTool: "",
+			wantErr:           true,
 		},
 		{
-			name: "get parent SCM URL",
-			pom: []PomXML{
-				{
-					GroupID:    "com.example",
-					ArtifactID: "child",
-					VersionID:  "1.0.0",
-					Parent: Parent{
-						GroupID:    "com.example.parent",
-						ArtifactID: "parent",
-						VersionID:  "1.0.0",
-					},
-				},
-				{
-					GroupID:    "com.example.parent",
-					ArtifactID: "parent",
-					VersionID:  "1.0.0",
-					SCMURL:     "https://github.com/example/parent",
-				},
-			},
-			targetPom: rebuild.Target{
-				Package: "com.example:child",
-				Version: "1.0.0",
-			},
-			expectedURL: "https://github.com/example/parent",
-			wantErr:     false,
+			name: "pom from src directory should be ignored",
+			repo: `
+            commits:
+              - id: initial-commit
+                files:
+                  src/test/resources/pom.xml: |
+                      <project></project>`,
+			expectedBuildTool: mavenBuildTool,
+			wantErr:           true,
 		},
-		{
-			name: "get URL of parent and not child",
-			pom: []PomXML{
-				{
-					GroupID:    "com.example",
-					ArtifactID: "child",
-					VersionID:  "1.0.0",
-					URL:        "https://example.com/child",
-					Parent: Parent{
-						GroupID:    "com.example.parent",
-						ArtifactID: "parent",
-						VersionID:  "1.0.0",
-					},
-				},
-				{
-					GroupID:    "com.example.parent",
-					ArtifactID: "parent",
-					VersionID:  "1.0.0",
-					URL:        "https://example.com/parent",
-				},
-			},
-			targetPom: rebuild.Target{
-				Package: "com.example:child",
-				Version: "1.0.0",
-			},
-			expectedURL: "https://example.com/parent",
-			wantErr:     false,
-		},
-		{
-			name: "should throw an error for no valid URL",
-			pom: []PomXML{
-				{
-					GroupID:    "com.example",
-					ArtifactID: "child",
-					VersionID:  "1.0.0",
-					Parent: Parent{
-						GroupID:    "com.example.parent",
-						ArtifactID: "parent",
-						VersionID:  "1.0.0",
-					},
-				},
-				{
-					GroupID:    "com.example.parent",
-					ArtifactID: "parent",
-					VersionID:  "1.0.0",
-				},
-			},
-			targetPom: rebuild.Target{
-				Package: "com.example:child",
-				Version: "1.0.0",
-			},
-			expectedURL: "",
-			wantErr:     true,
-		},
-	}
-	for _, tc := range testCases {
+	} {
 		t.Run(tc.name, func(t *testing.T) {
-			mockRegistry := &mockMavenRegistry{
-				artifactCoordinates: make(map[struct{ PackageName, VersionID, FileType string }][]byte),
-			}
-			for _, pom := range tc.pom {
-				addPomArtifact(mockRegistry, &pom)
-			}
-			mockMux := rebuild.RegistryMux{
-				Maven: mockRegistry,
-			}
-			got, err := Rebuilder{}.InferRepo(context.Background(), tc.targetPom, mockMux)
+			repo := must(gitxtest.CreateRepoFromYAML(tc.repo, nil))
+			head := must(repo.Head())
+			headCommit := must(repo.CommitObject(head.Hash()))
+			buildTool, err := inferBuildTool(headCommit)
 			if tc.wantErr {
 				if err == nil {
-					t.Fatalf("InferRepo() = %q, want error", got)
+					t.Fatalf("inferBuildTool() expected error but got none")
 				}
 			} else {
 				if err != nil {
-					t.Fatalf("InferRepo() error = %v", err)
+					t.Fatalf("inferBuildTool() error = %v", err)
 				}
-				if got != tc.expectedURL {
-					t.Errorf("InferRepo() = %q, want %q", got, tc.expectedURL)
+				if buildTool != tc.expectedBuildTool {
+					t.Errorf("inferBuildTool() = %v, want %v", buildTool, tc.expectedBuildTool)
 				}
 			}
-
 		})
 	}
 }
 
-func addPomArtifact(mavenRegistry *mockMavenRegistry, pom *PomXML) {
-	key := struct{ PackageName, VersionID, FileType string }{
-		PackageName: fmt.Sprintf("%s:%s", pom.GroupID, pom.ArtifactID),
-		VersionID:   pom.VersionID,
-		FileType:    maven.TypePOM,
+func must[T any](t T, err error) T {
+	if err != nil {
+		panic(err)
 	}
-	xml := "<project xmlns=\"http://maven.apache.org/POM/4.0.0\" xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\" xsi:schemaLocation=\"http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd\"><groupId>" + pom.GroupID + "</groupId><artifactId>" + pom.ArtifactID + "</artifactId><version>" + pom.VersionID + "</version>"
-	// Add parent if present
-	if pom.Parent.GroupID != "" && pom.Parent.ArtifactID != "" && pom.Parent.VersionID != "" {
-		xml += "<parent><groupId>" + pom.Parent.GroupID + "</groupId><artifactId>" + pom.Parent.ArtifactID + "</artifactId><version>" + pom.Parent.VersionID + "</version></parent>"
-	}
-	xml += "<url>" + pom.URL + "</url><scm><url>" + pom.SCMURL + "</url></scm></project>"
-	mavenRegistry.artifactCoordinates[key] = []byte(xml)
-}
-
-func TestFindPomXMLHeuristic(t *testing.T) {
-	testCases := []struct {
-		name        string
-		poms        []string
-		expectedPom string
-	}{
-		{
-			name:        "single pom.xml at root",
-			poms:        []string{"pom.xml"},
-			expectedPom: "pom.xml",
-		},
-		{
-			name:        "select pom.xml in subdir",
-			poms:        []string{"subdir/pom.xml"},
-			expectedPom: "subdir/pom.xml",
-		},
-		{
-			name:        "select a non-test resource pom.xml",
-			poms:        []string{"moduleA/pom.xml", "src/test/resource/pom.xml"},
-			expectedPom: "moduleA/pom.xml",
-		},
-		{
-			name:        "select pom.xml with basename match",
-			poms:        []string{"moduleA/test-pom.xml", "moduleB/pom.xml"},
-			expectedPom: "moduleB/pom.xml",
-		},
-		{
-			name:        "select pom.xml with src prefix in directory name",
-			poms:        []string{"srcmoduleA/pom.xml", "src/moduleB/pom.xml"},
-			expectedPom: "srcmoduleA/pom.xml",
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			fs := memfs.New()
-			repo, _ := git.Init(memory.NewStorage(), fs)
-			w, _ := repo.Worktree()
-			for _, pomPath := range tc.poms {
-				dir := path.Dir(pomPath)
-				if dir != "." {
-					fs.MkdirAll(dir, 0755)
-				}
-				// Write a minimal valid pom.xml content to memfs
-				f, err := fs.Create(pomPath)
-				if err != nil {
-					t.Fatalf("failed to create file: %v", err)
-				}
-				content := `<project><groupId>group</groupId><artifactId>artifact</artifactId><version>1.0.0</version></project>`
-				_, err = f.Write([]byte(content))
-				if err != nil {
-					t.Fatalf("failed to write file: %v", err)
-				}
-				f.Close()
-				w.Add(pomPath)
-			}
-			commitHash, _ := w.Commit("add multiple pom files", &git.CommitOptions{
-				Author: &object.Signature{
-					Name:  "Foo Bar",
-					Email: "foo@bar",
-					When:  time.Now(),
-				},
-			})
-			commit, _ := repo.CommitObject(commitHash)
-			_, name, err := findPomXML(commit, "group:artifact")
-			if err != nil {
-				t.Fatalf("findPomXML() error = %v", err)
-			}
-			if name != tc.expectedPom {
-				t.Errorf("findPomXML() = %q, want %q", name, tc.expectedPom)
-			}
-		})
-	}
+	return t
 }
